VALIDATOR-411 - decided to use parseInt instead of Long and forgot to change variable
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/validator/trunk@1781672 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/validator/routines/UrlValidator.java
@@ -415,7 +415,7 @@ public class UrlValidator implements Serializable {
             String port = authorityMatcher.group(PARSE_AUTHORITY_PORT);
             if (port != null && port.length() > 0) {
             	try {
-            		long iPort = Integer.parseInt(port);
+            		int iPort = Integer.parseInt(port);
             		if (iPort < 0 || iPort > 0xFFFF) {
             			return false;
             		}
